+++ libraries/test_session_current/src/main/java/androidx/media3/session/MockPlayer.java
@@ -1090,15 +1090,6 @@ public class MockPlayer implements Player {
   }
 
   /**
-   * @deprecated Use {@link #hasPreviousMediaItem()} instead.
-   */
-  @Deprecated
-  @Override
-  public boolean hasPreviousWindow() {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
    * @deprecated Use {@link #hasNextMediaItem()} instead.
    */
   @Deprecated
@@ -1114,15 +1105,6 @@ public class MockPlayer implements Player {
 
   @Override
   public boolean hasNextMediaItem() {
-    throw new UnsupportedOperationException();
-  }
-
-  /**
-   * @deprecated Use {@link #seekToPreviousMediaItem()} instead.
-   */
-  @Deprecated
-  @Override
-  public void previous() {
     throw new UnsupportedOperationException();
   }