+++ RELEASENOTES.md
@@ -57,8 +57,10 @@
     *   Use `HttpEngineDataSource` as the `HttpDataSource` when supported by the
         device.
 *   Remove deprecated symbols:
-    *   Remove deprecated `Player#hasPrevious` method. Use
-        `Player#hasPreviousMediaItem()` instead.
+    *   Remove deprecated `Player.hasPrevious`, `Player.hasPreviousWindow()`.
+        Use `Player.hasPreviousMediaItem()` instead.
+    *   Remove deprecated `Player.previous()`method. Use
+        `Player.seekToPreviousMediaItem()` instead.
     *   Remove deprecated `DrmSessionEventListener.onDrmSessionAcquired` method.
 
 ## 1.4